// src/types/enum.ts
export enum gsap_animation_sections {
  hero = 'gsap_hero_section',
  destination = 'gsap_destination_section',
  inspiration = 'gsap_inspiration_section',
  spotlight = 'gsap_spotlight_section',
  journeys = 'gsap_journeys_section',
  list_your_property = 'gsap_list_your_property_section',
  testimonial = 'gsap_testimonial_section',
  about = 'gsap_about_section',
  featured_in = 'gsap_featured_in_section'
}

export enum gsap_property_details_animation_sections {
  hero = 'gsap_pd_hero_section',
  property_overview = 'gsap_pd_overview',
  reserve = 'gsap_pd_reserve',
  location = 'gsap_pd_location',
  rooms = 'gsap_pd_rooms',
  amenities = 'gsap_pd_amenities',
  included_excluded_services = 'gsap_pd_included_excluded_services',
  home_truths = 'gsap_pd_home_truths',
  before_you_book = 'gsap_pd_before_you_book',
  stay_longer = 'gsap_pd_stay_longer',
  experiences = 'gsap_pd_experiences',
  story = 'gsap_pd_story',
  detailed_description = 'gsap_pd_detailed_description',
  guest_experience = 'gsap_pd_guest_experience',
  faq = 'gsap_pd_faq',
  similar_stays = 'gsap_pd_similar_stays',
  included_services_mobile = 'gsap_pd_included_services_mobile',
  available_services_mobile = 'gsap_pd_available_services_mobile',
  home_truths_mobile = 'gsap_pd_home_truths_mobile',
  story_mobile = 'gsap_pd_story_mobile',
  detailed_description_mobile = 'gsap_pd_detailed_description_mobile'
}